refactor: convertToDto() to fromEntity() #64
