+++ backend/src/main/java/com/example/backend/dashboard/dto/CaseResponse.java
@@ -1,12 +1,15 @@
 package com.example.backend.dashboard.dto;
-import com.example.backend.dashboard.domain.CaseEntity.CaseState;
-import com.example.backend.dashboard.domain.CaseEntity.CaseCategory;
+import com.example.backend.common.domain.CaseEntity;
+import com.example.backend.common.domain.CaseEntity.CaseState;
+import com.example.backend.common.domain.CaseEntity.CaseCategory;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.Builder;
 import lombok.Data;
 import java.time.LocalDateTime;
 
 @Data
+@Builder
 public class CaseResponse {
     private Integer id;
     private String police_name;
@@ -21,5 +24,20 @@ public class CaseResponse {
     private String video;
     private CaseState state;
     private String memo;
+
+    public static CaseResponse fromEntity(CaseEntity entity) {
+        return CaseResponse.builder()
+                .id(entity.getId())
+                .police_name(entity.getPolice().getName())
+                .police_rank(String.valueOf(entity.getPolice().getRank()))
+                .cctv_address(entity.getCctv().getAddress())
+                .date(entity.getDate())
+                .level(entity.getLevel())
+                .category(entity.getCategory())
+                .video(entity.getVideo())
+                .state(entity.getState())
+                .memo(entity.getMemo())
+                .build();
+    }
 
 }
